news-read/pyproject.toml
[project]
name = "news-read"
version = "0.1.0"
description = "财经新闻源工具：东方财富-财经早餐"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "akshare>=1.12.0",
  "pandas>=2.2.0",
  "firecrawl>=0.3.0",
  "openai>=1.68.0",
  "python-dateutil>=2.9.0",
]

[project.scripts]
news-read = "src.main:main"
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
